requirements.txt
sentence-transformers
faiss-cpu
langchain
streamlit
transformers
groq
numpy
